requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.37.1
django-bootstrap4==0.0.7
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.18.4
requests-oauthlib==1.0.0
simplejson==3.14.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
